+++ admin/tabs/cluster/objects/DetailPanel.ts
@@ -12,6 +12,7 @@ import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {Icon} from '@xh/hoist/icon';
 import {DetailModel} from './DetailModel';
 import './ClusterObjects.scss';
+import {fmtJson, timestampReplacer} from '@xh/hoist/format';
 
 export const detailPanel = hoistCmp.factory({
     model: creates(DetailModel),
@@ -42,7 +43,7 @@ export const detailPanel = hoistCmp.factory({
                         height: '100%',
                         showFullscreenButton: false,
                         editorProps: {lineNumbers: false},
-                        value: model.fmtStats(selectedAdminStats)
+                        value: fmtJson(selectedAdminStats, {replacer: timestampReplacer()})
                     })
                 })
             ]