+++ appcontainer/AppStateModel.ts
@@ -4,11 +4,10 @@
  *
  * Copyright © 2025 Extremely Heavy Industries Inc.
  */
-import {AppState, AppSuspendData, HoistModel, XH} from '@xh/hoist/core';
+import {AppState, AppSuspendData, HoistModel, PlainObject, XH} from '@xh/hoist/core';
 import {action, makeObservable, observable} from '@xh/hoist/mobx';
 import {Timer} from '@xh/hoist/utils/async';
-import {getClientDeviceInfo} from '@xh/hoist/utils/js';
-import {camelCase, isBoolean, isString, mapKeys} from 'lodash';
+import {camelCase, isBoolean, isString, mapKeys, pick} from 'lodash';
 
 /**
  * Support for Core Hoist Application state and loading.
@@ -24,8 +23,13 @@ export class AppStateModel extends HoistModel {
     suspendData: AppSuspendData;
     accessDeniedMessage: string = 'Access Denied';
 
+    /**
+     * Timestamp when the app first started loading, prior to even JS download/eval.
+     * Read from timestamp set on window within index.html.
+     */
+    readonly loadStarted: number = window['_xhLoadTimestamp'];
+
     private timings: Record<AppState, number> = {} as Record<AppState, number>;
-    private loadStarted: number = window['_xhLoadTimestamp']; // set in index.html
     private lastStateChangeTime: number = this.loadStarted;
 
     constructor() {
@@ -88,13 +92,14 @@ export class AppStateModel extends HoistModel {
                     timestamp: loadStarted,
                     elapsed: Date.now() - loadStarted - (timings.LOGIN_REQUIRED ?? 0),
                     data: {
-                        appVersion: XH.appVersion,
-                        appBuild: XH.appBuild,
-                        locationHref: window.location.href,
+                        clientId: XH.clientId,
+                        sessionId: XH.sessionId,
                         timings: mapKeys(timings, (v, k) => camelCase(k)),
-                        ...getClientDeviceInfo()
+                        clientHealth: XH.clientHealthService.getReport(),
+                        window: this.getWindowData(),
+                        screen: this.getScreenData()
                     },
-                    logData: ['appVersion', 'appBuild'],
+                    logData: ['clientId', 'sessionId'],
                     omit: !XH.appSpec.trackAppLoad
                 })
         });
@@ -109,5 +114,37 @@ export class AppStateModel extends HoistModel {
                 this.lastActivityMs = Date.now();
             });
         });
+    }
+
+    private getScreenData(): PlainObject {
+        const screen = window.screen as any;
+        if (!screen) return null;
+
+        const ret: PlainObject = pick(screen, [
+            'availWidth',
+            'availHeight',
+            'width',
+            'height',
+            'colorDepth',
+            'pixelDepth',
+            'availLeft',
+            'availTop'
+        ]);
+        if (screen.orientation) {
+            ret.orientation = pick(screen.orientation, ['angle', 'type']);
+        }
+        return ret;
+    }
+
+    private getWindowData(): PlainObject {
+        return pick(window, [
+            'devicePixelRatio',
+            'screenX',
+            'screenY',
+            'innerWidth',
+            'innerHeight',
+            'outerWidth',
+            'outerHeight'
+        ]);
     }
 }