+++ CHANGELOG.md
@@ -2,17 +2,52 @@
 
 ## v73.0.0-SNAPSHOT - unreleased
 
-* Support for reporting Client Version in Admin WebSockets tab.
+### ⚙️ Technical
+
 * Update the `FormModel` with configuration options to support persisting the form state - either the entire form
 or a provided subset of fields.
 
+## v72.4.0 - 2025-04-09
+
+### 🎁 New Features
+* New methods for formatting timestamps within nested JSON objects.  See `withFormattedTimestamps`
+  and `timestampReplacer` in the `@xh/hoist/format` package.
+* `ViewManagerConfig` takes new optional key `viewMenuItemFn` to allow ViewManager implementations
+  to customize the menu items for views in the view manager menu.
+
+### ⚙️ Technical
+* Added dedicated `ClientHealthService` for managing client health report. Additional enhancements
+  to health report to include information about web sockets, idle time, and page state.
+
+## v72.3.0 - 2025-04-08
+
+### 🎁 New Features
+
+* Added support for posting a "Client Health Report" track message on a configurable interval. This
+  message will include basic client information, and can be extended to include any other desired
+  data via `XH.clientHealthService.addSource()`. Enable by updating your app's
+  `xhActivityTrackingConfig` to include `clientHealthReport: {intervalMins: XXXX}`.
+* Enabled opt-in support for telemetry in `MsalClient`, leveraging hooks built-in to MSAL to collect
+  timing and success/failure count for all events emitted by the library.
+* Added the reported client app version as a column in the Admin Console WebSockets tab.
+
 ### 🐞 Bug Fixes
 
-* NULL check when accessing records in Store to prevent a thrown error
+* Improved fetch request tracking to include time spent loading headers as specified by application.
+
+### ⚙️ Technical
+
+* Update shape of returned `BrowserUtils.getClientDeviceInfo()` to nest several properties under new
+  top-level `window` key and report JS heap size / usage values under the `memory` block in MB.
+
+### 📚 Libraries
+
+* @azure/msal-browser `3.28 → 4.8.0`
 
 ## v72.2.0 - 2025-03-13
 
 ### 🎁 New Features
+
 * Modified `TabContainerModel` to make more methods `protected`, improving extensibility for
   advanced use-cases.
 * Enhanced `XH.reloadApp` with new argument to clear query parameters before loading.
@@ -25,8 +60,8 @@ or a provided subset of fields.
 
 ### 🐞 Bug Fixes
 
-* Prevent native browser context menu on Dash Canvas surfaces. It can hide the Dash Canvas custom
-  context menu when an app's `showBrowserContextMenu` flag is `true`.
+* Prevented native browser context menu from showing on `DashCanvas` surfaces and obscuring the
+  `DashCanvas` custom context menu.
 
 ## v72.1.0 - 2025-02-13